fix(arcane): P2.5 review -- tighten motor-clamp test, extract WorldToLocal + shared MakeSoft/kPi (DRY)
- PhysicsJointsTest.cpp: tighten motor torque-clamp CHECK from <50 to <2
  (a removed clamp would reach ~50 rad/s; the true clamped ceiling is ~0.0001)
- Joints.cpp: extract WorldToLocal() file-scope helper in anon namespace,
  replace both duplicate toLocal lambdas (Revolute/Weld + Wheel cases)
- SoftCoeffs.hpp (new): shared SoftCoeffs struct + MakeSoft() -- single
  canonical home for the b2MakeSoft formula; prevents drift as P3.x adds
  more soft consumers (CCD/Astra)
- PhysicsTypes.hpp: move kPi here as inline constexpr (shared across physics)
- SoftStep.cpp: drop local Soft/MakeSoft/kPi; include SoftCoeffs.hpp
- Joints.cpp: drop local Soft/MakeSoft/kPi; include SoftCoeffs.hpp
- Shapes.cpp, PhysicsShapesTest.cpp: remove now-ambiguous local kPi copies
  (Shapes.cpp used same value; test renamed local double to kPiD to preserve
  f64 analytic precision in hand-derived expected values)
SoftStep [solver] invariant numbers unchanged (15501/18 green).
[joints] numbers unchanged (641/10 green). [physics] 17820/121 green.

+++ Tests/src/PhysicsJointsTest.cpp
@@ -339,8 +339,8 @@ TEST_CASE("PhysicsJoints: motor respects the torque clamp", "[physics][joints]")
     w.Step(kStep);
     const Real a1 = w.GetAngle(disk);
     const Real rate = (a1 - a0) / kStep;
-    // The clamp keeps the rate FAR below the 50 rad/s target after one step.
-    CHECK(rate < Real(50));
+    // clamp keeps the one-step rate far below the 50 rad/s target; a removed clamp would blow past 2.
+    CHECK(rate < Real(2));
     CHECK(rate >= Real(0));
 }
